requirements.txt
llama-index==0.8.41
python-dotenv
faiss